# backend/app/schemas/auth.py
from pydantic import BaseModel, EmailStr

from .user import TokenPair, UserRead


class LoginRequest(BaseModel):
    email: EmailStr
    password: str


class LoginResponse(BaseModel):
    user: UserRead
    tokens: TokenPair


class RefreshRequest(BaseModel):
    refresh_token: str